chore: move to src folder
