--- tests/test_test_example.py
@@ -1,7 +1,0 @@
-
-
-from utils.test_example import add
-
-
-def test_answer():
-    assert add(3, 2) == 5
